# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Project",
        "description": "This is the API documentation for HomayTech test project.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "user_uuid",
                        "in": "query",
                        "description": "Filter by user UUID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filter by user email",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "card_number",
                        "in": "query",
                        "description": "Filter by card number",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 15
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Filter by user level",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2,
                                3,
                                4
                            ]
                        }
                    },
                    {
                        "name": "card_number_count",
                        "in": "query",
                        "description": "Filter by number of bank cards",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "user_uuid": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email"
                                            },
                                            "bank_cards": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            },
                                            "level": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}